bits of cleanup and rename donClick to handleClick
