+++ amundsen_application/static/js/config/config-utils.ts
@@ -1,6 +1,7 @@
 import AppConfig from 'config/config';
 import { BadgeStyle, BadgeStyleConfig } from 'config/config-types';
 import { TableMetadata } from 'interfaces/TableMetadata';
+import { convertText, CaseType } from 'utils/textUtils';
 
 import { FilterConfig, LinkConfig } from './config-types';
 
@@ -74,34 +75,6 @@ export function getFilterConfigByResource(
 ): FilterConfig {
   return AppConfig.resourceConfig[resourceType].filterCategories;
 }
-// TODO @allisonsuarez is this the best place for these functions?
-export enum CaseType {
-  LOWER_CASE = 'lowerCase',
-  SENTENCE_CASE = 'sentenceCase',
-  UPPER_CASE = 'upperCase',
-  TITLE_CASE = 'titleCase',
-}
-
-export function convertText(str = '', caseType: string): string {
-  switch (caseType) {
-    case CaseType.LOWER_CASE:
-      return str.toLowerCase();
-    case CaseType.SENTENCE_CASE:
-      return `${str.charAt(0).toUpperCase()}${str.slice(1).toLowerCase()}`;
-    case CaseType.UPPER_CASE:
-      return str.toUpperCase();
-    case CaseType.TITLE_CASE:
-      const splitStr = str.toLowerCase().split(' ');
-      for (let i = 0; i < splitStr.length; i++) {
-        splitStr[i] =
-          splitStr[i].charAt(0).toUpperCase() + splitStr[i].substring(1);
-      }
-      return splitStr.join(' ');
-    default:
-      return str;
-  }
-}
-
 
 /*
  * Given a badge name, this will return a badge style and a display name.
